I changed the library name to ACrypto because it was conflicting with ESP8266 experimental Crypto module and the Crypto library of Rhys weatherley. Changed the call of 'calc_size_n_pad' in aesSha256Hmac exemple to 'calcSizeAndPad'. Changed the includes accordingly. Changes to be committed: 	renamed:    Crypto.cpp -> ACrypto.cpp 	renamed:    Crypto.h -> ACrypto.h 	modified:   examples/aesSha256Hmac/aesSha256Hmac.ino 	modified:   examples/sha256/sha256.ino 	modified:   examples/sha256hmac/sha256hmac.ino 	modified:   library.properties
