+++ ACrypto.h
