+++ examples/aesSha256Hmac/aesSha256Hmac.ino
@@ -1,4 +1,4 @@
-#include <Crypto.h>             // AES 128 CBC with pkcs7, RNG, SHA256 and SHA256HMAC  
+#include <ACrypto.h>             // AES 128 CBC with pkcs7, RNG, SHA256 and SHA256HMAC  
 #include <base64.hpp>           // Base64 encode and decode without line breaks https://github.com/Densaugeo/base64_arduino
 
 /*
@@ -65,7 +65,7 @@ void setup() {
   AES aes(keyEncrypt, iv, AES::AES_MODE_128, AES::CIPHER_ENCRYPT);
 
   // create buffer for final message which will contain IV, encrypted message, and HMAC
-  int encryptedSize = aes.calc_size_n_pad(packetSize);
+  int encryptedSize = aes.calcSizeAndPad(packetSize);
   int ivEncryptedSize = encryptedSize + AES_KEY_LENGTH;
   int ivEncryptedHmacSize = ivEncryptedSize + SHA256HMAC_SIZE;
   uint8_t ivEncryptedHmac[ivEncryptedHmacSize];
