+++ examples/sha256/sha256.ino
@@ -1,4 +1,4 @@
-#include <Crypto.h>
+#include <ACrypto.h>
 
 /*
  * Compute the SHA256 hash of a message on an ESP8266
