+++ examples/sha256hmac/sha256hmac.ino
@@ -1,4 +1,4 @@
-#include <Crypto.h>
+#include <ACrypto.h>
 
 /*
  * Compute the SHA256HMAC of a message on an ESP8266
