+++ library.properties
@@ -1,4 +1,4 @@
-name=Crypto
+name=ACrypto
 version=1.0.0
 author=Chris Ellis
 maintainer=Chris Ellis
